# swagger/api-docs.yaml
openapi: 3.0.0
info:
  title: User Authentication API
  version: 1.0.0
  description: API for user login and logout functionality
servers:
  - url: http://localhost:3000/api # Replace with your actual base API URL
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Get all users
      description: Retrieves a list of all users in the system.
      responses:
        200:
          description: Successful retrieval of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error
    post:
      tags:
        - Users
      summary: Create a new user
      description: Creates a new user with the provided information.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                gmail: # Changed from 'gmail'
                  type: string
                  description: User's gmail address
                password:
                  type: string
                  description: User's password
                  format: password
                firstName:
                  type: string
                  description: User's first name
                lastName:
                  type: string
                  description: User's last name
                middleName:
                  type: string
                  description: User's middle name (optional)
                dateOfBirth:
                  type: string
                  description: User's date of birth (MM/DD/YYYY, optional)
                avatarImage:
                  type: string
                  format: binary
                  description: User's profile image (optional)
      responses:
        201:
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get a specific user
      description: Retrieves a single user based on their ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        200:
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCreated"
        404:
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: User not found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error
    put:
      tags:
        - Users
      summary: Update an user
      description: Updates an existing user with the provided information.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                gmail:
                  type: string
                  description: User's gmail address (optional)
                firstName:
                  type: string
                  description: User's first name (optional)
                lastName:
                  type: string
                  description: User's last name (optional)
                middleName:
                  type: string
                  description: User's middle name (optional)
                dateOfBirth:
                  type: string
                  description: User's date of birth (MM/DD/YYYY, optional)
                avatarImage:
                  type: string
                  format: binary
                  description: User's profile image (optional)
      responses:
        200:
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCreated"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Unauthorized
        404:
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: User not found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error
...
  /participants:
    post:
      tags:
        - Participants
      summary: Create a new participant
      description: Creates a new participant in a conversation with the provided information.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                conversationId:
                  type: integer
                  description: Conversation's ID
                gmail:
                  type: string
                  description: Participant's gmail address
      responses:
        201:
          description: Participant added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error
  /participants/{id}:
    get:
      tags:
        - Participants
      summary: Get participants in a specific conversation
      description: Retrieves an array of participants in conversation with conversationID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Conversation ID
      responses:
        200:
          description: Participants retrieved successfully
          content:
            application/json:
              schema:
                type: array
                $ref: "#/components/schemas/ParticipantInConversation"
        404:
          description: Participants not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Participants not found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error
    put:
      tags:
        - Participants
      summary: Update a participant
      description: Update a normal participant into admin of the conversation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Participant ID
      responses:
        200:
          description: Participant updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Unauthorized
        404:
          description: Participant not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Participant not found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error

    delete:
      tags:
        - Participants
      summary: Delete a participant
      description: Delete a participant from a conversation.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Participant ID
      responses:
        200:
          description: Participant deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Unauthorized
        404:
          description: Participant not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Participant not found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserCreated:
      type: object
      properties:
        id:
          type: integer
        gmail:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
        dateOfBirth:
          type: string
        avatarName:
          type: string
        disabled:
          type: boolean
    User:
      type: object
      properties:
        id:
          type: integer
        gmail:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
        dateOfBirth:
          type: string
        avatarName:
          type: string
        disabled:
          type: boolean
        conversation:
          type: array
          items:
            $ref: "#/components/schemas/ConversationGeneral"
        friends:
          type: array
          items:
            $ref: "#/components/schemas/Friend"
    ConversationGeneral:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        creatorId:
          type: integer
        imageName:
          type: string
        createdAt:
          format: date-time
        updatedAt:
          format: date-time
        deletedAt:
          format: date-time
...
    Participant:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        userId:
          type: integer
        isAdmin:
          type: boolean
        createdAt:
          format: date-time
        updatedAt:
          format: date-time
    ParticipantInConversation:
      type: object
      properties:
        id:
          type: integer
        gmail:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
        dateOfBirth:
          format: date-time
        avatarName:
          type: string
        disabled:
          type: boolean
        participant_details:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              conversationId:
                type: integer
              isAdmin:
                type: boolean
    Friend:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        friendId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...